feat: update refresh button component to make it reusable
