+++ components/TopActions.tsx
@@ -1,14 +1,32 @@
-import { FC } from 'react'
+import { FC, useState } from 'react'
 import { useRouter } from 'next/router'
 import { SearchBar, RefreshButton } from './'
-import { useRecoilState } from 'recoil'
-import { candyMachineSearchState } from 'lib/recoil-store/atoms'
+import { useRecoilState, useSetRecoilState } from 'recoil'
+import { candyMachineSearchState, candyMachinesState } from 'lib/recoil-store/atoms'
 import { Button } from '@primer/react'
 import { LinkExternalIcon } from '@primer/octicons-react'
+import { useWallet } from '@solana/wallet-adapter-react'
+import { useRPC } from 'hooks'
+import { fetchCandyMachineAccounts } from 'lib/utils'
 
 const TopActions: FC = () => {
     const [searchValue, setSearchValue] = useRecoilState(candyMachineSearchState)
     const { pathname } = useRouter()
+    const [isLoading, setIsLoading] = useState(false)
+    const { publicKey } = useWallet()
+    const { rpcEndpoint } = useRPC()
+    const setCandyMachines = useSetRecoilState(candyMachinesState)
+
+    const refreshCandyMachines = async () => {
+        setIsLoading(true)
+        try {
+            const candyMachines = await fetchCandyMachineAccounts(rpcEndpoint, publicKey!)
+            setCandyMachines(candyMachines)
+        } catch (e) {
+            console.error(e)
+        }
+        setIsLoading(false)
+    }
 
     return (
         <div className='d-flex flex-justify-end top-actions-bar d-flex flex-row'>
@@ -19,7 +37,7 @@ const TopActions: FC = () => {
                         setSearchValue={setSearchValue}
                         placeholderText='Search candy machine'
                     />
-                    <RefreshButton />
+                    <RefreshButton onClick={refreshCandyMachines} isLoading={isLoading} />
                 </>
             ) : (
                 <>
